oppdatert info om roboten

+++ ros_intro_ws/tek4030_turtlebot3-master/src/sample.cpp
@@ -42,8 +42,8 @@ int main (int argc, char **argv){
 
         //===========formlene=================================
 
-        int k = 0;
-        
+        double k = 0.0;
+
